+++ test/libsolidity/smtCheckerTests/functions/getters/string.sol
@@ -10,4 +10,6 @@ contract C {
 	}
 }
 // ----
-// Warning 6328: (211-257): CHC: Assertion violation happens here.
+// Warning 1218: (211-257): CHC: Error trying to invoke SMT solver.
+// Warning 6328: (211-257): CHC: Assertion violation might happen here.
+// Warning 4661: (211-257): BMC: Assertion violation happens here.