+++ test/libsolidity/smtCheckerTests/special/msg_sender_3.sol
@@ -1,0 +1,21 @@
+pragma experimental SMTChecker;
+
+contract C {
+	address owner;
+
+	constructor() {
+		owner = msg.sender;
+		assert(owner >= address(0)); // should hold
+	}
+}
+
+contract D {
+	address owner;
+
+	constructor() {
+		unchecked {
+			owner = msg.sender;
+			assert(owner >= address(0)); // should hold
+		}
+	}
+}